suggestions from review
