+++ STACpopulator/extensions/thredds.py
@@ -5,8 +5,10 @@ from pystac.extensions.base import (
     ExtensionManagementMixin,
     PropertiesExtension,
 )
-
-from STACpopulator.stac_utils import ServiceType, magpie_resource_link
+from pydantic import ConfigDict, field_validator, model_validator
+from STACpopulator.stac_utils import ServiceType, magpie_resource_link, ncattrs_to_bbox, ncattrs_to_geometry
+from STACpopulator.extensions.base import Helper, BaseSTAC
+from STACpopulator.extensions.datacube import DataCubeHelper
 
 T = TypeVar("T", pystac.Collection, pystac.Item)
 
@@ -110,7 +112,7 @@ class CollectionTHREDDSExtension(THREDDSExtension[pystac.Item]):
         return f"<CollectionTHREDDSExtension Collection id={self.obj.id}>"
 
 
-class THREDDSHelper:
+class THREDDSHelper(Helper):
     def __init__(self, access_urls: dict[str, str]):
         self.access_urls = {
             ServiceType.from_value(svc): url
@@ -138,6 +140,52 @@ class THREDDSHelper:
         ext = THREDDSExtension.ext(item, add_if_missing=add_if_missing)
         ext.apply(services=self.services, links=self.links)
         return item
+
+
+class THREDDSCatalogDataModel(BaseSTAC):
+    """Base class ingesting attributes loaded by `THREDDSLoader` and creating a STAC item.
+
+    This is meant to be subclassed for each extension.
+
+    It includes two validation mechanisms:
+     - pydantic validation using type hints, and
+     - json schema validation.
+    """
+    # Data from loader
+    data: dict
+
+    # Extensions classes
+    datacube: DataCubeHelper
+    thredds: THREDDSHelper
+
+    model_config = ConfigDict(populate_by_name=True, extra="ignore", arbitrary_types_allowed=True)
+
+    @classmethod
+    def from_data(cls, data):
+        """Instantiate class from data provided by THREDDS Loader.
+        """
+        # This is where we match the Loader's output to the STAC item and extensions inputs. If we had multiple
+        # loaders, that's probably the only thing that would be different between them.
+        return cls(data=data,
+                   start_datetime=data["groups"]["CFMetadata"]["attributes"]["time_coverage_start"],
+                   end_datetime=data["groups"]["CFMetadata"]["attributes"]["time_coverage_end"],
+                   geometry=ncattrs_to_geometry(data),
+                   bbox=ncattrs_to_bbox(data),
+                   )
+
+    @model_validator(mode="before")
+    @classmethod
+    def datacube_helper(cls, data):
+        """Instantiate the DataCubeHelper."""
+        data["datacube"] = DataCubeHelper(data['data'])
+        return data
+
+    @model_validator(mode="before")
+    @classmethod
+    def thredds_helper(cls, data):
+        """Instantiate the THREDDSHelper."""
+        data["thredds"] = THREDDSHelper(data['data']["access_urls"])
+        return data
 
 
 # TODO: Validate services links exist ?
