--- STACpopulator/implementations/CORDEX-CMIP6_Ouranos/__init__.py
@@ -1,4 +1,0 @@
-from .add_CORDEX6 import add_parser_args, runner
-
-__all__ = ["add_parser_args", "runner"]
-
