+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/__init__.py
@@ -1,0 +1,4 @@
+from STACpopulator.implementations.CORDEXCMIP6_Ouranos .add_CORDEX6 import add_parser_args, runner
+
+__all__ = ["add_parser_args", "runner"]
+
