+++ tests/test_cordex.py
@@ -1,3 +1,4 @@
+import json
 from STACpopulator.extensions.cordex6 import Cordex6DataModel, Cordex6DataModelNcML
 
 
@@ -5,7 +6,7 @@ def get_first_item_attrs(url):
     import requests
     from siphon.catalog import TDSCatalog
     import xncml
-    from STACpopulator.stac_utils import numpy_to_python_datatypes
+    from STACpopulator.stac_utils import np
 
     cat = TDSCatalog(url)
 
@@ -13,14 +14,29 @@ def get_first_item_attrs(url):
         for item_name, ds in cat.datasets.items():
             r = requests.get(ds.access_urls["NCML"])
             attrs = xncml.Dataset.from_text(r.text).to_cf_dict()
-            attrs["attributes"] = numpy_to_python_datatypes(attrs["attributes"])
             attrs["access_urls"] = ds.access_urls
-            return attrs
+            return numpy_to_python_datatypes(attrs)
+
+
+def make_test_data():
+    """Fetches attribute data from the PAVICS THREDDS catalog and stores it in the data/ directory as a json."""
+    # Raw CORDEX data
+    url = "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/catalog/birdhouse/disk2/ouranos/CORDEX/CMIP6/DD/NAM-12/OURANOS/MPI-ESM1-2-LR/ssp370/r1i1p1f1/CRCM5/v1-r1/day/tas/v20231208/catalog.html"
+    attrs = get_first_item_attrs(url)
+    with open("data/cordex6_raw.json", "w") as f:
+        json.dump(attrs, f)
+
+    # NcML CORDEX data
+    url = "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/catalog/datasets/simulations/RCM-CMIP6/CORDEX/NAM-12/day/catalog.html"
+    attrs = get_first_item_attrs(url)
+    with open("data/cordex6_ncml.json", "w") as f:
+        json.dump(attrs, f)
+
 
 
 def test_item_raw():
-    url = "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/catalog/birdhouse/disk2/ouranos/CORDEX/CMIP6/DD/NAM-12/OURANOS/MPI-ESM1-2-LR/ssp370/r1i1p1f1/CRCM5/v1-r1/day/tas/v20231208/catalog.html"
-    attrs = get_first_item_attrs(url)
+
+    attrs = json.load(open("tests/data/cordex6_raw.json"))
     model = Cordex6DataModel.from_data(attrs)
     item = model.stac_item()
 
@@ -29,9 +45,8 @@ def test_item_raw():
 
 
 def test_item_ncml():
-    url = "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/catalog/datasets/simulations/RCM-CMIP6/CORDEX/NAM-12/day/catalog.html"
 
-    attrs = get_first_item_attrs(url)
+    attrs = json.load(open("tests/data/cordex6_ncml.json"))
     model = Cordex6DataModelNcML.from_data(attrs)
     item = model.stac_item()
